Add: refactor data module with peslab submodule. Adds peslab data quality checks (adapted from Tom's old old old code)
